requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-groq>=0.1.3
langchain-core>=0.2.0
python-dotenv>=1.0.0